+++ app/application/services/get_trajectories_service.go
@@ -6,10 +6,10 @@ import (
 
 	"github.com/kajiLabTeam/walking-information-storage-server-api/app/infrastructure"
 	"github.com/kajiLabTeam/walking-information-storage-server-api/app/infrastructure/repository"
-	"github.com/kajiLabTeam/walking-information-storage-server-api/app/presentation/controllers"
+	dto_presenttion "github.com/kajiLabTeam/walking-information-storage-server-api/app/presentation/dto"
 )
 
-func GetTrajectoriesService(floorID string) (*controllers.GetTrajectoriesResponse, error) {
+func GetTrajectoriesService(floorID string) (*dto_presenttion.GetTrajectoriesResponse, error) {
 	db, err := infrastructure.ConnectionDB()
 	if err != nil {
 		log.Fatal("データベース接続エラー:", err)
@@ -18,14 +18,14 @@ func GetTrajectoriesService(floorID string) (*controllers.GetTrajectoriesRespons
 
 	//floor_id(フロアID)を元に軌跡ID(trajectry_id)を取得
 	//Todo:floorIDに置き換える
-	trajectories, err := repository.GetTrajectoriesIDByFloorID(db, "01F8VYXK67BGC1F9RP1E4S9YTV")
+	trajectories, err := repository.GetTrajectoriesIDByFloorID(db, floorID)
 	if err != nil {
 		log.Printf("軌跡IDの取得エラー: %w", err)
 	}
 	fmt.Println("trajectories")
 	fmt.Println(trajectories)
 
-	resTrajectories := []controllers.Trajectory{}
+	resTrajectories := []dto_presenttion.Trajectory{}
 
 	for _, trajectory := range trajectories {
 		//軌跡ID(trajectry_id)に紐付いた推定座標(estimated_positions)/正解座標(correct_positions)を取得
@@ -39,12 +39,12 @@ func GetTrajectoriesService(floorID string) (*controllers.GetTrajectoriesRespons
 			log.Printf("正解軌跡の取得エラー: %w", err)
 		}
 		// 推定座標(estimatedPositions)を配列にして、推定軌跡(estimatedTrajectories)に変形
-		estimated := []controllers.Position{}
-		correct := []controllers.Position{}
+		estimated := []dto_presenttion.Position{}
+		correct := []dto_presenttion.Position{}
 
 		//推定座標estimatedPositionをPosition型に変更する
 		for _, estimatedPosition := range estimatedPositions {
-			position := controllers.Position{
+			position := dto_presenttion.Position{
 				ID:       estimatedPosition.ID,
 				X:        float32(estimatedPosition.X),
 				Y:        float32(estimatedPosition.Y),
@@ -52,11 +52,11 @@ func GetTrajectoriesService(floorID string) (*controllers.GetTrajectoriesRespons
 			}
 
 			estimated = append(estimated, position)
-			fmt.Printf("estimated: %v\n", estimated)
+			// fmt.Printf("estimated: %v\n", estimated)
 		}
 		//推定座標estimatedPositionをPosition型に変更する
 		for _, correctPosition := range correctPositions {
-			position := controllers.Position{
+			position := dto_presenttion.Position{
 				ID:       correctPosition.ID,
 				X:        float32(correctPosition.X),
 				Y:        float32(correctPosition.Y),
@@ -64,17 +64,17 @@ func GetTrajectoriesService(floorID string) (*controllers.GetTrajectoriesRespons
 			}
 
 			correct = append(correct, position)
-			fmt.Printf("correct: %v\n", correct)
+			// fmt.Printf("correct: %v\n", correct)
 		}
 
-		resTrajectory := controllers.Trajectory{
+		resTrajectory := dto_presenttion.Trajectory{
 			ID:        trajectory.ID,
 			Estimated: estimated,
 			Correct:   correct,
 		}
 
 		resTrajectories = append(resTrajectories, resTrajectory)
-		fmt.Printf("resTrajectories: %v\n", resTrajectories)
+		// fmt.Printf("resTrajectories: %v\n", resTrajectories)
 	}
 
 	//正解座標(CorrectPositions)を配列にして、正解軌跡(CorrectTrajectories)にする
@@ -83,12 +83,12 @@ func GetTrajectoriesService(floorID string) (*controllers.GetTrajectoriesRespons
 		log.Printf("データベースのクローズに失敗: %v", err)
 	}
 
-	floor := controllers.Floor{
+	floor := dto_presenttion.Floor{
 		ID:          "1",
 		MapImageURL: "署名付きURL",
 	}
 
-	trajectoriesResponse := controllers.GetTrajectoriesResponse{
+	trajectoriesResponse := dto_presenttion.GetTrajectoriesResponse{
 		Floor:        floor,
 		Trajectories: resTrajectories,
 	}
